Initial Porject Setup:
We have decided to fork from my version to something edina can get behind,
I have updated references to my GitHub repository to the forked edina
version. I also fixed the indent size in .editorconfig as well as updating
the package.json file with the new name and URL as well as the existing
contributors.