IssueID #2 - Configure whether to run proxy API:
So far we have only wanted to run the proxy API server, for front-end
development purposes. Now we want to integrate the front-end/back-end
code. We need to be able to pass proxied requests to the real back-end
not just to the proxy API server.

Now, if you configure start to be false, the system will just forward
requests to the proxy you define. If start = true, then it will also
run the proxy task, which should setup the proxy API server.y